refactor: split lifecycle.rs into lifecycle/mod.rs + lifecycle/setup.rs to satisfy source-length budget
